Refactor: scope and packages for httpclient in core
