+++ core/src/main/scalajvm/sttp/client3/quick.scala
@@ -2,5 +2,4 @@ package sttp.client3
 
 object quick extends SttpApi {
   lazy val backend: SttpBackend[Identity, Any] = HttpClientSyncBackend()
-  lazy val urlConnectionBackend: SttpBackend[Identity, Any] = HttpURLConnectionBackend()
 }
