+++ core/src/test/scalajvm/sttp/client3/HttpClientFutureWebSocketTest.scala
@@ -1,7 +1,6 @@
-package sttp.client3.httpclient
+package sttp.client3
 
 import sttp.capabilities.WebSockets
-import sttp.client3.{HttpClientFutureBackend, SttpBackend}
 import sttp.client3.testing.ConvertToFuture
 import sttp.client3.testing.websocket.{WebSocketConcurrentTest, WebSocketTest}
 import sttp.monad.{FutureMonad, MonadError}
